rofl-relayer/src/rofl_relayer/utils/contract_utility.py
import json
from pathlib import Path

from eth_account import Account
from eth_account.signers.local import LocalAccount
from sapphirepy import sapphire
from web3 import Web3
from web3.middleware import SignAndSendRawMiddlewareBuilder


class ContractUtility:
    """
    Utility for contract interaction and ABI loading.
    
    Can be used in two modes:
    1. Full mode: Initialize with RPC URL and secret for signing transactions
    2. Read-only mode: Initialize with RPC URL only for building unsigned transactions
    """

    def __init__(self, rpc_url: str, secret: str = ""):
        """
        Initialize the ContractUtility.
        
        Args:
            rpc_url: RPC URL for the network (required)
            secret: Private key for signing transactions (optional - if not provided, read-only mode)
        """
        if not rpc_url:
            raise ValueError("RPC URL is required")
            
        self.rpc_url = rpc_url
        
        # Always create Web3 instance with RPC
        self.w3 = Web3(Web3.HTTPProvider(self.rpc_url))
        
        # Add signing middleware only if secret is provided
        if secret:
            self._add_signing_middleware(secret)

    def _add_signing_middleware(self, secret: str) -> None:
        """
        Add signing middleware to the existing Web3 instance.
        
        Args:
            secret: Private key for signing transactions
        """
        if not secret:
            raise ValueError("Private key is required for signing transactions")

        account: LocalAccount = Account.from_key(secret)
        self.w3.middleware_onion.add(SignAndSendRawMiddlewareBuilder.build(account))
        # self.w3 = sapphire.wrap(self.w3, account)
        self.w3.eth.default_account = account.address

    def get_contract_abi(self, contract_name: str) -> list:
        """Fetches ABI of the given contract from the contracts folder"""
        contract_path = (
            Path(__file__).parent.parent.parent.parent
            / "contracts"
            / f"{contract_name}.json"
        ).resolve()
        
        with contract_path.open() as file:
            contract_data = json.load(file)

        return contract_data["abi"]
