rofl-relayer/pyproject.toml
[project]
name = "rofl-relayer"
version = "0.1.0"
description = "ROFL Relayer for blockchain bridge operations"
requires-python = ">=3.12"
dependencies = [
    "oasis-sapphire-py>=0.4.0",
    "aiohttp",
    "web3",
    "rlp>=3.0.0",
    "trie>=2.0.0",
    "httpx",
    "eth-utils>=5.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
    "pytest",
    "pytest-asyncio",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rofl_relayer"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.12+
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
select = [
    "E4",      # pycodestyle errors
    "E7",
    "E9",
    "F",       # pyflakes
    "I",       # isort
    "UP",      # pyupgrade
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "ICN",     # flake8-import-conventions
    "PIE",     # flake8-pie
    "PYI",     # flake8-pyi
    "RSE",     # flake8-raise
    "RUF",     # Ruff-specific rules
    "SIM",     # flake8-simplify
    "T20",     # flake8-print
]

ignore = [
    "E501",    # line too long, handled by formatter
    "B008",    # do not perform function calls in argument defaults
    "C901",    # too complex
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["rofl_relayer"]
